refactor: dev server without babel
