+++ package.json
@@ -63,7 +63,6 @@
     "next": "9.0.6",
     "next-routes": "1.4.2",
     "node-fetch": "2.6.0",
-    "nodemailer": "6.3.0",
     "nprogress": "0.2.0",
     "nuka-carousel": "4.5.12",
     "payment": "2.3.0",
@@ -114,7 +113,7 @@
   },
   "scripts": {
     "start": "node dist/server",
-    "dev": "babel-node server",
+    "dev": "node server",
     "build": "npm-run-all build:updates build:clean build:next build:server",
     "build:clean": "shx rm -rf dist .next",
     "build:langs": "npm run langs:update",
