+++ server/index.js
@@ -1,26 +1,26 @@
-import '../env';
+require('../env');
 
-import path from 'path';
-import http from 'http';
-import express from 'express';
-import next from 'next';
-import accepts from 'accepts';
-import cookieParser from 'cookie-parser';
-import cloudflareIps from 'cloudflare-ip/ips.json';
+const http = require('http');
+const path = require('path');
 
-import routes from './routes';
-import { loggerMiddleware, logger } from './logger';
-import { getLocaleDataScript, getMessages, languages } from './intl';
+const next = require('next');
+const express = require('express');
+const cookieParser = require('cookie-parser');
+const cloudflareIps = require('cloudflare-ip/ips.json');
+
+const intl = require('./intl');
+const logger = require('./logger');
+const loggerMiddleware = require('./logger-middleware');
+const routes = require('./routes');
 
 const server = express();
 
 server.set('trust proxy', ['loopback', 'linklocal', 'uniquelocal'].concat(cloudflareIps));
 
 const env = process.env.NODE_ENV;
-const dev = env === 'development' || env === 'docker';
+const dev = env === 'development';
 
 const app = next({ dev, dir: path.dirname(__dirname) });
-server.next = app;
 
 const port = process.env.PORT;
 
@@ -29,21 +29,7 @@ app.prepare().then(() => {
 
   server.use(cookieParser());
 
-  server.use((req, res, next) => {
-    if (req.query.language && languages.includes(req.query.language)) {
-      // Detect language as query string in the URL
-      req.language = req.query.language;
-    } else if (req.cookies.language && languages.includes(req.cookies.language)) {
-      // Detect language in Cookie
-      req.language = req.cookies.language;
-    }
-
-    req.locale = req.language || accepts(req).language(languages) || 'en';
-    logger.debug('url %s locale %s', req.url, req.locale);
-    req.localeDataScript = getLocaleDataScript(req.locale);
-    req.messages = getMessages(req.locale);
-    next();
-  });
+  server.use(intl.middleware());
 
   server.use(routes(server, app));
   server.use(loggerMiddleware.errorLogger);
