--- server/middlewares.js
@@ -1,6 +1,0 @@
-export const maxAge = (maxAge = 60) => {
-  return (req, res, next) => {
-    res.set('Cache-Control', `public, max-age=${maxAge}`);
-    next();
-  };
-};
